refactor: score를 처리하는 로직을 TeamScore 클래스에 구현
